versioning of PCInstance (#1299)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1299

# What:
Implement versioning:
I modified `load_schema` so we can safely delete attributes in PCInstance now.
More examples can be found here: D37841149

# Why:

Since we always have this conversion between json and Python object, the key of versioning is `load_schema`: we can convert an old version json into a new version Python object.

Safe operations:
1. delete attributes
2. add default value for existing attributes
3. add new attribute with default value
4. when modify a default value / change optional type
  ok: my_str: str = "Hello, world" -> my_str: str
  ok: my_int: Optional[int] = None -> my_int: Optional[int]
  ok: my_int: Optional[int] = None -> my_int: int = 4
  not ok: my_int: Optional[int] = None -> my_int: int

Some case which will fail:
1. rename attributes
2. add a new required attribute

Same for
  not ok: my_int: Optional[int] = None -> my_int: int

I feel the key point is: If there are some attributes missing in an old version, how do we handle these attributes? Do we want a crash so we know the error? or do we want to give it a default value so the computation run can continue?
1 -> crash: we know this immediately, and have to rerun
2 -> continue with default value: a default value can lead to wired output and may take much unnecessary efforts to debug. Or even worse, the output seems fine, but we cannot trust it..

Reviewed By: gorel

Differential Revision: D37894104

fbshipit-source-id: 367e3b5571cfe2e335568917da0c1188b48d3257