Cohort ID support in data processing stage for partner data (#1326)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1326

In Private Lift, the partner may want to include an identifier for each event in the partner data to identify which cohort a user belongs to. This is simply an integer and is useful for the advertiser to calculate cohort lift. `cohort_id` had been briefly supported in the past, but due to problems in how it was handled in the PL flow, it has since been removed. We now want to permanently add back support for this field. This is a requested feature from multiple advertisers, and is currently a blocker in onboarding for those who need it.

In adding back support, the main thing is ensuring compatibility between the data processing stage and the following compute stage. The data processing stage prepares the input for the compute stage by aggregating the conversion entries for a single user into one entry. In combining the events for a single user into one row, it aggregates various fields into arrays. By default, the data processing stage will aggregate all columns provided that are not the `id_` column. This works perfectly for all expected fields except for `cohort_id`, since the PL compute stage expects only one cohort id to be provided for each entry. The PL run will fail.

In theory, a user should only belong to one cohort across all events. This means that we can dedupe the cohort ids gathered for one user into just one value to comply with the expectations of the compute stage. Thus, we can explicitly remove `cohort_id` from the columns to aggregate and just return back one `cohort_id` per user. Everything should flow smoothly from there.

More context here: https://fburl.com/gdoc/1fe9htyj

Reviewed By: chualynn, anthonyzhang25

Differential Revision: D38013630

fbshipit-source-id: a7afa52faf88bff092e9a1a6db75e9542264393b