Test files to be used by ShardCombiner for initial accumulate tests
Summary:
# context
This commit adds a test file for a simple metric accumulation so that we know the basic logic of ShardCombiner is correct.

Basically there are 2 shards that are Xor split into partner and publisher.

ShardCombiner is supposed to add 100 + 90 as the initial test :D

Also adds expected output for previously created xor-ss shards.

# changes in the diff
- expected output json
- inputs plaintext
- inputs xor-ss
- plaintext_attribution_out.json shards (so that we can test whether AggMetrics is created and opening it to parties yields correct json outputs).

Differential Revision: D38021614

fbshipit-source-id: d423c03ad2e28ba7bc4d3b2b91d882faf42e58be