Add tests for lift with cohorts and breakdowns (#1356)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1356

[PCF][PL] Add tests for lift with cohorts and breakdowns to verify the correctness of the implementation using framework, against a plaintext implementation that does not use any cryptography.
We currently already have tests against a fixed set of correctness cases as well as random test inputs, and for this task we will add tests for missing functionalities for the random test inputs. In particular, the input data generator for the random test inputs does not include partner cohorts or publisher breakdowns, so we will first add that, then add tests to verify correctness.

Reviewed By: chualynn

Differential Revision: D38078384

fbshipit-source-id: 655f339a9716fab87363722b711ffe2a8c7080c0