Adding local stage flow to fix automated e2e tests. (#1365)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1365

# Context
Caleb is working on adding the automated test for pcf2 lift, however we are facing an issue where the test times out due to an underlying error in aggregation stage.

# Error
Aggregation stage is looking for input in compute stage i.e. in PCF 1.0 lift, while we are running with PCF2.0 lift and the files are places in pcf2_lift_stage.

# Fix
Added PCF2LiftLocalStageFlow to aggregation service, so that it picks up the right input path for the local tests.

Reviewed By: jrodal98

Differential Revision: D38213046

fbshipit-source-id: b1bdb8c2ec1b33321d404ac529a945e7bbb51911