Add Ad Id compression logic to the attribution game
Summary:
Based on the flag, add a logic to retrieve the original Ad Ids from touchpoint data and create a compression map of the original Ad Id with the compressed Ad ID, replace the original Ad Id with compressed Ad Ids in the attribution logic and output generation.

The adid is in plaintext rather than secret share for XOR input encryption. For the plaintext scheduler, this means that both parties adid are the same, so when running this line, the adid will be 0, because they are treated as secret share rather than plaintext. If input encryption is XOR, we should store secret shared ad ids instead of plaintext. Thus when doing the ad id compression, the adId should also be in plaintext regardless of input encryption.

Reviewed By: ajinkya-ghonge

Differential Revision: D38223281

fbshipit-source-id: 46ab9852ef8b534cf9234f3d33bd61ec306dac2d