ShardCombinerApp add test file for AdObj format
Summary:
# context
The gameTest (previous commits) mainly tests accumulation logic on toy dataset and LiftMetrics checks. For the sake of completeness I'm adding one of the test case that was removed due because it had kanonymity written over it.

NOTE: I'm not changing the name of the file for the sake of consistency with the old tests and the naming here https://fb.quip.com/IUHDApxKEAli#CHLACA6N8rP

# Changes in the diff
- test files moved from previous directory :)

Reviewed By: anthonyzhang25

Differential Revision: D38250136

fbshipit-source-id: 1064869b803b04f91021c2dafedf4ef9da622764