Aggregate non-Id columns for PL publisher in Id Combiner (#1409)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1409

# Context
Currently, PL cannot handle multiple rows with the same private-id. Therefore, we need to develop de-duplication method for PL. For single-key, we are de-duplicating in the PL Dataswarm. For multi-key, the de-duplication is difficult problem to solve.

In [the doc](https://fb.workplace.com/groups/779808116318681/posts/894761421490016), we listed and discussed possible approaches to de-duplicate for multi-key data. As a result of discussion, we decided to develop de-duplication logic in PID PREPARE. Please see [the doc](https://docs.google.com/document/d/1DlEuQkWGp2_STqjK5ebAZL_bTKL0nc6YDggQpSL56XE/edit?usp=sharing) for the details of this approach.

# Changes
In this diff, we have done the following.
1. Created `aggregateLiftNonIdColumns()` to aggregate the columns as [the PL Dataswarm aggregates duplicates](https://fburl.com/code/1oi5o255).
2. Added additional arg `isPublisherLift` in `idSwapMultiKey()`, which is set to false as default. If the flag is on, we call `aggregateLiftNonIdColumns()` to aggregate columns
3. In the `LiftIdSpineFileCombiner()`, we pass `idSwapMultiKey(isPublisherLift=true)` when it is publisher run.
4. Added unit test in `IdSwapMultiKeyTest.cpp` and `LiftIdSpineFileCombinerTest.cpp` to cover all the code path.

Reviewed By: ajinkya-ghonge

Differential Revision: D38435930

fbshipit-source-id: 507fd2bc342886367488c3c5707f5e5bc6047e4b