Add optional parameter to target specific platform when building FBPCS images
Summary:
Docker images are built with a default target depending on the local system building the image. This can result in inconsistent builds across different build systems.

As of Docker Engine 1.40+, Docker support a `--platform` argument which can be used to specify an image target.

This change adds an optional parameter, `-p PLATFORM` to the `build-docker.sh` script so that build systems can reliably build against a specific target.

Reviewed By: musebc

Differential Revision: D38624511

fbshipit-source-id: d93d836f437c9bf0ea4369b705115492dffb22f0