Add optional parameter to target specific platform when building FBPCS images (#1436)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1436

Docker images are built with a default target depending on the local system building the image. This can result in inconsistent builds across different build systems.

As of Docker Engine 1.40+, Docker support a `--platform` argument which can be used to specify an image target.

This change adds an optional parameter, `-p PLATFORM` to the `build-docker.sh` script so that build systems can reliably build against a specific target.

Reviewed By: musebc

Differential Revision: D38624511

fbshipit-source-id: 9fe207f14171c90b4f37861de342b71c79d883ee