Remove partial dependency in PCSContainerService (#1428)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1428

X-link: https://github.com/facebookresearch/fbpcp/pull/405

We decided that validate_container_definitions should not be a public API in container service. This diff is to remove partial dependency in PCSContainerService

The validation logic has been moved in D38629555, to ensure the backward compatibility, we firstly changed the function body to ``pass``.

Reviewed By: joe1234wu

Differential Revision: D38604420

fbshipit-source-id: 9b7b370ee66b483841a40eca64a70ab1a0f3f204