Install missing build dependency for the FBPCS pce_deployment image
Summary:
An issue was encountered that prevented build of the `pce_deployment` image against the aarch64/arm64 platform.

Building the Docker image against aarch64 fails due to some missing build dependencies: `gcc`, `python3-dev`. This is because a project dependency, the `psutil` wheel, is not distributed with the aarch64 build and instead it is compiled from source (see [Wheel support for linux aarch64 · Issue #1782 · giampaolo/psutil · GitHub](https://github.com/giampaolo/psutil/issues/1782)).

This change updates the corresponding Dockerfile to install the dependencies required to compile the wheel.

Reviewed By: ajaybhargavb

Differential Revision: D38541052

fbshipit-source-id: 9ce75bd3016de062ba601649849840f85d881a6e