Add a cron GitHub action to build the Onedocker image every hour of the work day. (#1440)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1440

## Background
Currently we build the onedocker image on every commit to FBPCS. This helps us catch most of the changes that affect the process. This does not include changes to PID and FBPCP though.

## This change
This commit adds a cron action that will build the OneDocker image every hour during the work day (8 AM - 5 PM PDT).

Reviewed By: gitfish77

Differential Revision: D38709566

fbshipit-source-id: 19cbd20d86c986cf1f28fa81988e1afadd294e9e