Adding string formatter to keep all the formatted string in one location (#1437)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1437

**What**
`download_logs.py` file has the python code which helps in downloading container logs and uploading logs to AWS S3. This change refactors on how we create formatted strings for file paths

**Why**
With the scattered format strings in the file, it gets difficult to manage them. This change puts all the format strings managed from one function so that it gets easier to make changes in the changes.

**Context**
https://docs.google.com/document/d/1cMD9IQ8uF6MS1PRvOh8cr-1el7MO952SjUanAzAGJzw/edit?usp=sharing

Reviewed By: marksliva

Differential Revision: D38679369

fbshipit-source-id: d52336b99a373012aeedea88be67afbcc6bd2712