Write the mapping compressed adId to original adId json file to S3
Summary:
Write a json of mapping compressed Ad ID to original Ad Id to S3, based on the flag, along with pcf2_attribution output.

In Aggregation games, we compress the Ad Id so that we can get most performance gain out of write only ORAM operation. Now we are sending the secret share value of Ad ID to the aggregation game, we will not be able to compress the Ad ID in the aggregation game, so we need to do the same in the attribution game. Based on the previous diff: D38223281, we have compressed the original Ad Id to compressed ad Id, now we need to store the mapping of compressed ad Id to original ad Id so that the mapping will be read on the shard aggregation game.

Reviewed By: chualynn

Differential Revision: D38372047

fbshipit-source-id: 85ae72486c04cfe1f36f6b51a68ba995fc96daac