fix S288948: must not query container state from OneDocker service while the container is already stopped
Summary:
The issue:
In AWS, a container's state can no longer be queried after the container had stopped for more than 1-hour. In this case, OneDocker cannot return the container state.
The fix:
Skip querying any container from OneDocker when the container is in one of the terminal states.
Similar to past fix in D34410047.

Differential Revision: D38799272

fbshipit-source-id: 61ffbc8edfe6ef6877c1d6e1911e24d52d9c7c9d