fix S288948: must not query container state from OneDocker service while the container is already stopped (#1470)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1470

The issue:
In AWS, a container's state can no longer be queried after the container had stopped for more than 1-hour. In this case, OneDocker cannot return the container state.
The fix:
Skip querying any container from OneDocker when the container is in one of the terminal states.
Similar to past fix in D34410047.

Reviewed By: joe1234wu

Differential Revision: D38799272

fbshipit-source-id: 80a74c3eb4fdb8761a942de693767eb26b331d19