Add a cron GitHub action to build the Coordinator image every hour during the work day. (#1453)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1453

## Background
Currently we build the coordinator and onedocker images on every commit to FBPCS. This helps us catch most of the changes that affect the process. This does not include changes to PID and FBPCP though. To properly test any changes that would happen, we need both the OneDocker and Coordinators built and to test with them.

## This change
This commit adds a cron action that will build the OneDocker image every hour during the work day (8 AM - 5 PM PDT).

Reviewed By: gitfish77

Differential Revision: D38750486

fbshipit-source-id: 3f92788eb709ef0065745b14b6ee21519e77bb2a