ad-hoc validation [1/n] (#1474)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1474

Ad-hoc validation is a way to dynamically allowing a user to validate their infrastructure using PCE validator.   Previously PCE validator was only used at the end of a deployment , but quickly we realized the need for ad-hoc validation because advertiser wanted to make sure their infrastructure is correct before then start a run.

What is planned as part of the overall project ?

1. Changes in java service layer which handles deployment. (this diff)
2. Changes in kotlin GraphQl service layer which proxies the api from frontend.
3. Changes in frontend .
4. Refactor old code to incorporate some newest way. (mostly in java sevrice layer to begin with)
5. Test & further improve and optimize.

This diff(#1 above):

1. Adds a couple of generic shell command handler classes.
2. exposes an API under V2
3. Changes in docker images to include some scripts.
4. Some script level changes to incorporate ad-hoc validation. Which will get little more refactored in #5

Differential Revision: D38822819

fbshipit-source-id: 0217b677d57bdb4b1fe7920ad7a867c3fe512653