RC E2E test for PL on PCF 2.0 (#1504)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1504

We cannot use the existing RC tests for PL on PCF 2.0. This is because we will not be able to test the existing prod path i.e. regression testing. Thus I am creating a new PL test specifically for RC tier.
This test will use the business - Rivendell Grocery, which we will add to GK of features that have made it to RC. This way we would be able to test new RC features as well as do regression testing.

Differential Revision: D38970556

fbshipit-source-id: e264120c601adec786b21c18e259d1c5a47d2182