Fix Hotfix action
Summary:
## Background
Right now the hotfix action is broken because the hotfix branch already exists on the origin and is not in the same order as the main branch we are creating the new hotfix branch from, causing the following error on push

https://pxl.cl/2bRlg

## This change
This change updates the action to force push the new hotfix branch to overwrite the old one. This will allow for the overwrite to the new hotfix branch and we can continue the hotfix process from there.

Differential Revision: D39056131

fbshipit-source-id: f616d4e0f3da0acef6ffdb22c46d02b423817527