Enable PID MR stage Flow using feature gate (#1515)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1515

Context:
As we plan to pilot run pid mr, we created a new stage flow that runs PID. In order to enable this feature for a limited pilot run advertisers we add a feature gate (GK - private_attribution_with_mr_pid) that will allow specific users. This feature gate needs to be used to pick specific stage flow based on the game type.

As we plan to run preliminary tests for attribution and Lift. We use the same GK to pick the appropriate stageflow based on the game type. We may add different GK for rollout.

This diff:
Uses PCSFeatures to select the appropriate MR pid  stage flow class based on the game type.

Reviewed By: wenqingren

Differential Revision: D38811938

fbshipit-source-id: 43b08188eaf8054c660740d98f4e2c4eb634a038