added regression test for partner output.txt logs in Github one_command_runner_test workflow (#1492)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1492

Output.txt logs are redirected to /tmp/output.txt in the AWS build machine, and the logs are always printed into the workflow message upon success or failure.
When runner failed, the workflow will abort.

Why this change is required:
In the migration to Bolt runner, some logs required by log analyzer are changed or missing. Adding this regression test will be able to detect such breaking change.

Reviewed By: ankushksingh, joe1234wu, ajitfb

Differential Revision: D38849701

fbshipit-source-id: b7c9b0bad4c51f483dca84643877a4325d939cac