Don't copy InputProcessor members into Attributor and Aggregator members (#1499)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1499

Currently these classes are copying some values from the InputProcessor and storing them as local member variables. This is a bad idea as it creates extra copies of the data and in the future could diverge on implementation.

For my immediate usage in the next diffs I plan to convert the `inputProcessor` constructor variable into a `std::unique_ptr` and when we call `std::move()` on it, we lose all these values. Either we first copy them and then std::move the inputProcessor, or we just use the source of truth (the inputProcessor) in the business logic directly which is more cleaner.

Reviewed By: chualynn

Differential Revision: D39002510

fbshipit-source-id: db477437eeb0582d9599c3b20e7670491d8fc893