Input Processor interface (#1520)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1520

Adding an interface for the InputProcessor which has methods that will be used by the Attributor and Aggregator classes.

The current InputProcessor is instantiated using an [InputData](https://fburl.com/code/2x99njmu) object which reads from the local filesystem one of the input csv's from the reshard stage. The idea behind this interface is that we don't actually care how we load those MPC types into the InputProcessor. We can add an InputProcessor which reads from a file of secret shares (computed in another stage), as well as a InputProcessor which runs the UDP algorithm on the the plaintext data. This is exactly what we will do in later steps.

Reviewed By: RuiyuZhu

Differential Revision: D39003554

fbshipit-source-id: a19b9e55dda280fb291228e934036fabb6955574