Adding aws specific exceptions in aws cloud library (#1521)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1521

**What**
Adding specific exceptions in `aws_cloud.py` for aws error.

**Why**
We had bare exceptions in `aws_cloud.py` and had a BE task to fix those. Fixing that in this diff. With the new data infra logging changes coming in this file it is best to address these issues now

**Context**
https://docs.google.com/document/d/1cMD9IQ8uF6MS1PRvOh8cr-1el7MO952SjUanAzAGJzw/edit?usp=sharing

Differential Revision: D39046168

fbshipit-source-id: e62c35a4c600d84dfb439d9bfa4fe86da770282c