Update Input processor interface to use data struct (#1528)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1528

Followup from D39003554 (https://github.com/facebookresearch/fbpcs/commit/0da7ae38de98f0501a5515186267f770f962e4c3). This diff is quite a handful as it refactors a bunch of PCF2 PL code to work with a new data struct called `LiftGameProcessedData`.

Previous approach was to directly call getters for each member variable in `InputProcessor`, however this approach is prone to creating lots of copies as data is not returned by reference.

This API for InputProcessor now just includes a

`getLiftGameProcessedData` which returns a const ref to the MPC'ified input data. We can add serialization / deserialization functionality in a future diff to make the secret share input processor simple to reason about.

- Removed all getters from IInputProcessor
- Moved members variables into a single member struct `LiftGameProcessedData`
- Replaced all references in code to getters to use the new struct.

It's a lot of code changes but mostly non-functional.

Reviewed By: mdas7

Differential Revision: D39178354

fbshipit-source-id: e1b790b56e6d3b58ba03a0ef15a59272c6dcf6b9