Ad-hoc validation [4/n]
Summary:
Ad-hoc validation is a way to dynamically allowing a user to validate their infrastructure using PCE validator.   Previously PCE validator was only used at the end of a deployment , but quickly we realized the need for ad-hoc validation because advertiser wanted to make sure their infrastructure is correct before then start a run.

What is planned as part of the overall project ?

1. Changes in java service layer which handles deployment. (D38822819 (https://github.com/facebookresearch/fbpcs/commit/067c3e4fd45eb184e9bb178f4e819f9e7266a77e))
2. Changes in kotlin GraphQl service layer which proxies the api from frontend. (D38880576)
3. Some refactoring and finally enabling PCE validator to read credentials from environment variables (this diff)
4. Changes in frontend .
4. Change in kotlin level to support integration
5. Test & further improve and optimize.

This diff(#3 above):

1. enables PCE validator to read tokens from environment variables, thus enabling using session token during development.
2. Removed set -e which was allowing the shell script to exit directly

Differential Revision: D39407298

fbshipit-source-id: 6eea38e099d2887d1edddccf3477c50e3213a777