Get Kinesis firehose details and upload logs on S3
Summary:
**What**
This diff is part of the data infra pipeline logging project. AWS kinesis firehose is used to process events in the pipeline. In this diff we get gather error logs and configs and upload on S3 along with container and deployment logs

This diff uploads 2 logs:
1. kinesis error logs
2. kinesis configs

Error logs will be used to check if any errors were encountered in processing the events. Configs are used to check status of the kinesis and if logging is enabled in the kinesis.

**Why**
Right now we don't capture errors in kinesis. This diff will help with that. Additionally, it will help in fetching kinesis configs, to help us in debugging in case any misconfiguration is causing the issue.

**Context**
https://docs.google.com/document/d/1cMD9IQ8uF6MS1PRvOh8cr-1el7MO952SjUanAzAGJzw/edit?usp=sharing

Differential Revision: D39286548

fbshipit-source-id: 1dbb00823730209e31813dd28d78cf5aa26ffded