Serialize and Deserialize LiftGameProcessedData array values (#1581)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1581

This diff adds serialization and deserialization of the following values in LiftGameProcessedData:

  SecTimestamp<schedulerId> opportunityTimestamps;
  SecBit<schedulerId> isValidOpportunityTimestamp;
  SecBit<schedulerId> anyValidPurchaseTimestamp;
  SecBit<schedulerId> testReach;

As each row has a single value, it will end up in the CSV as a single value per row.

In the deserialization step we collect all the rows into a vector for each value to be loaded into the MPC type.

This diff also handles the numRows_ logic as well as making the secret shares output work with the reshard logic (https://fburl.com/code/y35xbauq)

Reviewed By: RuiyuZhu

Differential Revision: D39547913

fbshipit-source-id: 9b037833ba88a6ade9853955f1f3a2b882a4b6f0