Serialize and deserialize constant values in LiftGameProcessedData (#1580)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1580

Implements partial serialization of constants in LiftGameProcessedData as added in D39322161. Unimplemented values will just have dummy data.

  uint32_t numPartnerCohorts;
  uint32_t numPublisherBreakdowns;
  uint32_t numGroups;
  uint32_t numTestGroups;
  uint8_t valueBits;
  uint8_t valueSquaredBits;

These are written as a single line to a CSV file. The sharder will not need to touch these parameters, and they will be used in each shard of the UDP data.

Reviewed By: RuiyuZhu

Differential Revision: D39443227

fbshipit-source-id: fa1eb3f08a84e2d1410ff5d225d017a4fad700f2