Fix table name issue in the generated policy
Summary:
When the events stream path was changed to 'events_data' in D39047627 (https://github.com/facebookresearch/fbpcs/commit/b1101514321ef35b5a3cc1c3dee5e8025814f0f0), it was
not possible to quickly test that the generated policy was compatible with the
new table name generated from the new path. However this turned out to be an
issue when testing the resource-based auth flow so I have fixed it in this
stack.

Differential Revision: D39640460

fbshipit-source-id: 908a0885697834159a586477cac465c4660afb21