Add PL UDP App with Dummy implementation (#1614)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1614

This diff adds the PL UDP App + a Dummy game that performs the plaintext input processing and outputs an IInputProcessor. It is set up with the same inputs as used in the existing PL binary.

Currently it runs the existing InputProcessor code and outputs the results to the two files provided to the App. In a future diff we will implement an InputProcessor which performs UDP to compress the outputs.

Reviewed By: chualynn

Differential Revision: D39710063

fbshipit-source-id: 777158f35f276de135d9df11abe76740821d3495