Add spine combiner binary
Summary:
# Context
[Private ID DFCA one-pager](https://docs.google.com/document/d/1mwImSRCljqU66uXWJS7jL-dQOrYAR4aBZ3bpJ8quiJw)

Advertisers want to be able to target their customers on our platform in a privacy-safe manner. We utilise Private ID with the existing DFCA solution to provide this.
We're adding a new game type `PRIVATE_ID_DFCA` whose inputs for both parties are csvs with the columns `user_pii, user_id_{party}`, and the output is a csv of `user_id_partner, user_id_publisher` which can then be fed to our existing DFCA APIs.
After running the existing PID stages, we need two new binaries to get the output in its final form:
- private_id_dfca_spine_combiner
- private_id_dfca_aggregator

# This Diff
Adds the private_id_dfca_spine_combiner binary to the existing ID_SPINE_COMBINER stage.
The code is 99% the same as the PL/PA spine combiners.
We assume there are no duplicate PIIs in the data file and there is at most 1 match for every ID combination.

# This Stack
- **Spine combiner binary**
- Aggregation stage config
- Aggregation stage binary
- Stage flow
- Game type

Differential Revision: D39579146

fbshipit-source-id: ac1158160495d247d82ec8a07402ef5e1b85ca1d