Add Validation for cohort id during PC run. (#1646)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1646

 cohort ids are optionally provided by the advertiser in partner data. Cohort ids should start from 0 and increment by 1.
However, there is currently no validation in place to make sure that partner side data follows this format.
We received a SEV - S290240, where this format was not followed and as result Lift stage faced a timeout issue.

Thus, in this diff, adding a validation so that partner side data follows the cohort id format.

The long term solution would be cohort ID compression (similar to Ad Id Compression.)

Reviewed By: jrodal98, anthonyzhang25

Differential Revision: D39800839

fbshipit-source-id: 23711416ab5ed71dee7ae9a5e598e183dce5851d