fbpkg test failure due to compressionMapping json.
Summary:
During one of the unit tests for reformatted attribution format. We are creating a json file which gets created in root folder in this case (will be created in S3 in prod).
Due to this file creation, we are gettting error in fbpkg creation.
In this diff, removing the file if it exists after the test run.

Differential Revision: D39994394

fbshipit-source-id: f01ae2525a18bd47e1d76d9de5b8262d5f4b860b