fbpkg test failure due to compressionMapping json. (#1672)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1672

During one of the unit tests for reformatted attribution format. We are creating a json file which gets created in root folder in this case (will be created in S3 in prod).
Due to this file creation, we are gettting error in fbpkg creation.
In this diff, removing the file if it exists after the test run.

Reviewed By: yanglu-fb

Differential Revision: D39994394

fbshipit-source-id: 23d5e5b22364b28716a285ccf344f2ee3c7822a9