Update FBPCS build pipeline to explicitly target amd64 when building images
Summary:
From task:

In a recent change, we added an argument to the FBPCS build script build-docker.sh to allow the caller to explicitly pass a platform to target (-p PLATFORM) when building an image.  Previously, we were implicitly targeting the platform that build was running on: amd64. However, we always require amd64 images to run in our cloud deployment.
This task is to update our build system to pass this new argument and explicitly target amd64 or linux/amd64. This will require our build system to run Docker Engine 1.40+.

Reviewed By: musebc, gitfish77

Differential Revision: D40026703

fbshipit-source-id: 0174659b0e33a7d4872c922de24fb7276019dd54