Adding cli changes to download data infra logs
Summary:
**What**
This diff is part of the data infra pipeline logging project. In this diff `download_logs_cli` is changed to add options to enable data infra pipeline logging.

**Why**
`download_logs_cli` is called from computation UI when the computation run is completed. Right now it fetches only container logs and uploads on S3 is a zipped folder. With this change it will upload data infra pipeline logs as well.

**Context**
https://docs.google.com/document/d/1cMD9IQ8uF6MS1PRvOh8cr-1el7MO952SjUanAzAGJzw/edit?usp=sharing

Differential Revision: D40152846

fbshipit-source-id: 66cae2cc39ce9891fb237bed78f218ee2a3b7d4a