Adding cli changes to download data infra logs (#1694)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1694

**What**
This diff is part of the data infra pipeline logging project. In this diff `download_logs_cli` is changed to add options to enable data infra pipeline logging.

**Why**
`download_logs_cli` is called from computation UI when the computation run is completed. Right now it fetches only container logs and uploads on S3 is a zipped folder. With this change it will upload data infra pipeline logs as well.

**Context**
https://docs.google.com/document/d/1cMD9IQ8uF6MS1PRvOh8cr-1el7MO952SjUanAzAGJzw/edit?usp=sharing

Reviewed By: ramesc

Differential Revision: D40152846

fbshipit-source-id: 4be62103edd725815e1faa8d11d48afa2da78abd