Create Mock test for DP Noise in DotproductGame (#1714)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1714

This diff mocks the value of DP noise to test that dp noise is added correctly to the dotproduct. The two parameters are delta and eps. DP noise is generated as gaussian noise and passed to the dotproduct game.

This is done by creating a new class MockDotproductGame and overriding the return value of its member function generateDPNoise().

Differential Revision: D40219980

fbshipit-source-id: 8eefa10c890e553782fcac219bf6268e3fd1fad1