Split the "Publish Onedocker" workflow into build and test jobs (#1713)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1713

## Background
We are in the process of trying do various projects that require us to modify the way we use our GitHub self hosted runners. As part of that, we are splitting the E2E test runners and the build runners. I have already added a new "build only" runner that we can use to execute our builds on. This commit updates the Github actions that need to build the binaries to use that runner only.

Reviewed By: gorel

Differential Revision: D40241459

fbshipit-source-id: a065f57581fae7831dddd9263b54fc8606753d1e