Add binary of UDP library (#1715)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1715

This diff added the binary of UDP library as an emp game under ``//fbpcs/emp_games/data_processing/unified_data_process``.
 1. adds `UdpProcessGame.h` and `UdpProcessGame_impl.h` to implement a game for udp adapter and data processor.
2. adds `UdpProcessGameFactory.h` to create `UdpProcessGame`.
3. adds `UdpProcessApp.h` and `UdpProcessApp_impl.h` to wrap the `UdpProcessGame`.
4. adds a test file to test the `UdpProcessApp`

This game mainly expect four input parameters: ``party`` to indicate whether running the game as publisher or partner; ``row_number`` represents the size of ``unionMap`` from PID stage. ``row_size`` reoresents the number of bytes in a row of metadata. ``intersection`` represents the number of instersection in the ``unionMap``.

The ``UdpProcessApp`` will generate some random data based on the above input parameters and feed them into the ``UdpProcessGame``. In more detail, the ``dataGeneration()`` methods will generate a ``unionMap`` to simulate the results from PID stage and ``metadata`` to simulate each party's meta data.

``unionMap`` will be an integer vector with size ``row_number``. On publisher side, the first ``intersection`` rows in the ``unionMap`` will be positive indexes and the rest will be 0. On the partner side, the first ``intersection`` rows in the ``unionMap`` will be positive indexes and the rest will be -1. In this way, the top ``intersection`` rows in the ``unionMap`` represents the intersection.

Reviewed By: robotal

Differential Revision: D39939986

fbshipit-source-id: e7fe3209e251e9bc78b3604e6f505b8454af2764