Fix the ability to build binaries based on the local FBPCF image. (#1724)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1724

There was a bug in the code that pinned the FBPCF version which made it impossible to test with local FBPCF changes. This commit adds a new version flag which allows you to specify the FBPCF version, and defaults it to latest for developers who are testing. This commit also updates the build and release system to pin the FBPCF version for releases.

Reviewed By: adshastri

Differential Revision: D40285359

fbshipit-source-id: 177caec1eb748df1cfbfc196927420e9ea3800b9