Bump keys around container starts in RunBinaryBaseService
Summary:
This is useful for various SEVs we've seen around container issues - it doesn't really do much right now, but it's a start. The problem is that most of the "interesting" stages are in MPCService, which we can't integrate with since that would create a circular dependency of `fbpcs -> fbpcp -> fbpcs`.

To really fix this, we need to move MPCService into fbpcs.

Differential Revision: D40358018

fbshipit-source-id: 4bcfd8a6d026d37fce6266b9acc1d34a7d7fff5e