Update E2E test to run on the specified E2E test runners
Summary: This commit updates all E2E tests to only run on the E2E test runners. These runners will stay in the fb-ads-crypto-dev cloud account where all the configuration is set up and maintained. This will enable me to move the build runner to the meta-pci-aws-dev cloud project so it is under PSI On-Call ownership for debugging issues. It will also allow us to easily scale the test runners to support parallel tests.

Differential Revision: D40241460

fbshipit-source-id: a939d78c50d572f31d784817821db6556f611f3f