Update E2E test to run on the specified E2E test runners (#1751)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1751

This commit updates all E2E tests to only run on the E2E test runners. These runners will stay in the fb-ads-crypto-dev cloud account where all the configuration is set up and maintained. This will enable me to move the build runner to the meta-pci-aws-dev cloud project so it is under PSI On-Call ownership for debugging issues. It will also allow us to easily scale the test runners to support parallel tests.

Reviewed By: marksliva

Differential Revision: D40241460

fbshipit-source-id: a0c5c1e8295c89cdabf3c8f7027c71f24843baf9