Add num MPC container mutation (#1755)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1755

# Context
In this diff, we are adding logic to mutate num MPC container based on PID spine file size. Please see [the doc](https://fb.workplace.com/groups/pidmatchingxfn/posts/518728833410109) for more details on the context.

# Details
We implemented a logic to read a metrics file and mutate num MPC containers in the beginning of ID combiner. If metrics file does not exist, it would raise exception and fail immediately.

Later, if there are discrepancy in the shard size on both side, it would fail in the check of num MPC containers and number of server IPs at https://fburl.com/code/v749paw6.

Differential Revision: D40399873

fbshipit-source-id: 6f7fcabffbc7f433fa7ff8075b0f6c1d2a866fe0