Add build scripts for the Data Processing and EMP game images. (#1738)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1738

## Background
Now that we can reliably re-build the emp-games and data processing docker images on each commit, we can move the build to being more of a push methodology rather than a pull one at conveyor build time. This will help us to save time on each bundle creation.

## This commit
This change adds new jobs that run on every push to main. They build the EMP Games and Data Processing docker images and publish them to the GitHub container registry with the latest and sha tags. This is the first step in allowing us to cut out a 30 minute build process on every bundle we create.

Reviewed By: marksliva

Differential Revision: D40384201

fbshipit-source-id: e6aefe3af11c66816cdc0c1fa6f01a3ba6199ee1