skeleton of the certificate provider
Summary:
This is the first of a stack of diffs on getting and passing the certificates from the stage services to onedocker container.

This diff:
Provide a unified interface to retrieve TLS certificates for different parties. In stage services, we will be calling get_tls_certificates on the specific provider class. For publisher, we expect to get root_cert, server_cert, and private key from PCInstanceRepo; for partner, we expect to get root_cert from the parameters that got passed in during initialization.

Upcoming diffs:
call get_tls_certificates in stage services and pass it down to the onedocker container.

Reviewed By: danbunnell

Differential Revision: D40369181

fbshipit-source-id: 31e0699d8882ece4303a648cd8772cc255c11de3