Adding logic to get features in bolt_runner. (#1777)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1777

We need to add gatekeeping to - cancel publisher stage in bolt_runner.
While creating the publisher instance, bolt_runner gets a list of feature list back and saves it to the instance. In this diff, adding a function to bolt_client to get the features for partner side as well as publisher side.

Reviewed By: jrodal98

Differential Revision: D40529997

fbshipit-source-id: 48841c9d50228ecbbdde5607b7131fae58a19aae