Remove Clean-Up step in Publish Onedocker (#1783)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1783

## Background
As part of our goals to parallelize the end to end tests, we need to create a pool of test runners. This is not currently safe because if two instances of "Publish OneDocker" are running at the same time, the second instance would remove the output files while the first instance is still running.

## The fix
To solve this issue, we are just removing the clean up step that was run as part of the Publish OneDocker task. ECS containers are now cleaned up automatically withing 12 hours and I will add and S3 bucket rule to automatically clean up old output files that are more than 30 days old.

Reviewed By: joe1234wu

Differential Revision: D40558109

fbshipit-source-id: eae218e6aeb5bb619fdf54bb913dd9bde8a4c719